requirements.txt
mysqlclient
django
djangorestframework
djoser
django-cors-headers
